apps/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
python-dotenv>=1.0.0
httpx>=0.27.0
Pillow>=10.0.0
json-repair>=0.30.0
python-multipart>=0.0.9
mutagen>=1.47.0
pydub>=0.25.1
